doc more funcs
